Move Intersect and Union typeclasses to logic package
