+++ core/src/main/scala/com/rallyhealth/vapors/core/logic/Intersect.scala
@@ -1,4 +1,4 @@
-package com.rallyhealth.vapors.core.data
+package com.rallyhealth.vapors.core.logic
 
 trait Intersect[A] {
